+++ apps/storybook/.storybook/orderlyProvider.tsx
@@ -17,6 +17,7 @@ import {
   PhantomWalletAdapter,
   SolflareWalletAdapter,
 } from "@solana/wallet-adapter-wallets";
+import { customChains } from "./customChains";
 
 const network = WalletAdapterNetwork.Devnet;
 
@@ -80,6 +81,7 @@ const OrderlyProvider: FC<{ children: ReactNode }> = (props) => {
         //     variant: "text",
         //   },
         // }}
+        customChains={customChains as any}
       >
         {props.children}
       </OrderlyAppProvider>