+++ apps/storybook/tailwind.config.cjs
@@ -19,6 +19,7 @@ module.exports = withTV({
     "../../packages/markets/src/**/*.{ts,js,tsx,jsx,mdx}",
     "../../packages/chart/src/**/*.{ts,js,tsx,jsx,mdx}",
     "../../packages/ui-transfer/src/**/*.{ts,js,tsx,jsx,mdx}",
+    "../../packages/ui-cross-deposit/src/**/*.{ts,js,tsx,jsx,mdx}",
     "../../packages/ui-share/src/**/*.{ts,js,tsx,jsx,mdx}",
     "../../packages/ui-tradingview/src/**/*.{ts,js,tsx,jsx,mdx}",
     "../../packages/wallet-connector/src/**/*.{ts,js,tsx,jsx,mdx}",