+++ packages/ui-positions/src/components/liquidation/mobile/liquidationCell.script.tsx
@@ -1,0 +1,14 @@
+import { API } from "@orderly.network/types";
+
+export const useLiquidationCellScript = (props: {
+    item: API.Liquidation;
+    onSymbolChange?: (symbol: API.Symbol) => void;
+    index: number;
+}) => {
+
+    return {
+        ...props
+    };
+};
+
+export type LiquidationCellState = ReturnType<typeof useLiquidationCellScript>;