+++ packages/ui-positions/src/components/positionHistory/mobile/index.ts
@@ -1,0 +1,3 @@
+export { PositionHistoryCell } from "./positionHistoryCell.ui";
+export { PositionHistoryCellWidget } from "./positionHistoryCell.widget";
+export { usePositionHistoryCellScript } from "./positionHistoryCell.script";