+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,125 @@
 # @orderly.network/cli
 
+## 0.3.18-developV2-1.14
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.14
+
+## 0.3.18-developV2-1.13
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.13
+
+## 0.3.18-developV2-1.12
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.12
+
+## 0.3.18-developV2-1.11
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.11
+
+## 0.3.18-developV2-1.10
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.10
+
+## 0.3.18-developV2-1.9
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.9
+
+## 0.3.18-developV2-1.8
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.8
+
+## 0.3.18-developV2-1.7
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.7
+
+## 0.3.18-developV2-1.6
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.6
+
+## 0.3.18-developV2-1.5
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.5
+
+## 0.3.18-developV2-1.4
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.4
+
+## 0.3.18-developV2-1.3
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.3
+
+## 0.3.18-developV2-1.2
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.2
+
+## 0.3.18-developV2-1.1
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.1
+
+## 0.3.18-developV2-1.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.1-developV2-1.0
+
 ## 0.3.17
 
 ### Patch Changes