+++ packages/create-orderly-app/CHANGELOG.md
@@ -1,5 +1,95 @@
 # create-orderly-app
 
+## 1.5.18-developV2-1.14
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.13
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.12
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.11
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.10
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.9
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.8
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.7
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.6
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.5
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.4
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.3
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.2
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.1
+
+### Patch Changes
+
+- update
+
+## 1.5.18-developV2-1.0
+
+### Patch Changes
+
+- update
+
 ## 1.5.17
 
 ### Patch Changes