+++ packages/ui-leverage/CHANGELOG.md
@@ -1,5 +1,155 @@
 # @orderly.network/ui-leverage
 
+## 2.0.1-developV2-1.14
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.14
+  - @orderly.network/types@2.0.1-developV2-1.14
+  - @orderly.network/ui@2.0.1-developV2-1.14
+
+## 2.0.1-developV2-1.13
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.13
+  - @orderly.network/types@2.0.1-developV2-1.13
+  - @orderly.network/ui@2.0.1-developV2-1.13
+
+## 2.0.1-developV2-1.12
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.12
+  - @orderly.network/types@2.0.1-developV2-1.12
+  - @orderly.network/ui@2.0.1-developV2-1.12
+
+## 2.0.1-developV2-1.11
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.11
+  - @orderly.network/types@2.0.1-developV2-1.11
+  - @orderly.network/ui@2.0.1-developV2-1.11
+
+## 2.0.1-developV2-1.10
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.10
+  - @orderly.network/types@2.0.1-developV2-1.10
+  - @orderly.network/ui@2.0.1-developV2-1.10
+
+## 2.0.1-developV2-1.9
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.9
+  - @orderly.network/types@2.0.1-developV2-1.9
+  - @orderly.network/ui@2.0.1-developV2-1.9
+
+## 2.0.1-developV2-1.8
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.8
+  - @orderly.network/types@2.0.1-developV2-1.8
+  - @orderly.network/ui@2.0.1-developV2-1.8
+
+## 2.0.1-developV2-1.7
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.7
+  - @orderly.network/types@2.0.1-developV2-1.7
+  - @orderly.network/ui@2.0.1-developV2-1.7
+
+## 2.0.1-developV2-1.6
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.6
+  - @orderly.network/types@2.0.1-developV2-1.6
+  - @orderly.network/ui@2.0.1-developV2-1.6
+
+## 2.0.1-developV2-1.5
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.5
+  - @orderly.network/types@2.0.1-developV2-1.5
+  - @orderly.network/ui@2.0.1-developV2-1.5
+
+## 2.0.1-developV2-1.4
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.4
+  - @orderly.network/types@2.0.1-developV2-1.4
+  - @orderly.network/ui@2.0.1-developV2-1.4
+
+## 2.0.1-developV2-1.3
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.3
+  - @orderly.network/types@2.0.1-developV2-1.3
+  - @orderly.network/ui@2.0.1-developV2-1.3
+
+## 2.0.1-developV2-1.2
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.2
+  - @orderly.network/types@2.0.1-developV2-1.2
+  - @orderly.network/ui@2.0.1-developV2-1.2
+
+## 2.0.1-developV2-1.1
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.1
+  - @orderly.network/types@2.0.1-developV2-1.1
+  - @orderly.network/ui@2.0.1-developV2-1.1
+
+## 2.0.1-developV2-1.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@2.0.1-developV2-1.0
+  - @orderly.network/types@2.0.1-developV2-1.0
+  - @orderly.network/ui@2.0.1-developV2-1.0
+
 ## 2.0.0
 
 ### Major Changes